+++ Changes.md
@@ -4,6 +4,7 @@ Version 3.6.0.4
   1. Use module templates as cannot override the Mustache engine with our own due to module locallib rendering
      preventing overriding.
   2. Add favicon setting to general settings.  Provides the ability to use your own custom favicon.
+  3. Add login page background settings from Essential.
 
 Version 3.6.0.3
   1. Add H5P custom CSS setting from Essential.